Add default implementation for reconnect delay callback
